+++ core/entities/src/commonMain/kotlin/geo/Route.kt
@@ -2,10 +2,8 @@
  * Copyright (c) 2024 Pointyware. Use of this software is governed by the GPL-3.0 license.
  */
 
-package org.pointyware.xyz.core.entities.ride
+package org.pointyware.xyz.core.entities.geo
 
-import org.pointyware.xyz.core.entities.geo.Length
-import org.pointyware.xyz.core.entities.geo.Location
 import kotlin.time.Duration
 
 /**
